exercise 4.15-4.21

+++ Part4/bloglist/controllers/blog.js
@@ -1,25 +1,79 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
+
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
   
   blogRouter.post('/', async(request, response) => {
     const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid or missing' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+ 
     if (body.title && body.url) {
-      const blog = new Blog(body)
+      const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user.id
+      })
+
+      if (blog.likes === undefined) {
+        Object.assign(blog, { likes: 0 });
+      }
+
       const savedBlog = await blog.save()
+      
+      user.blog = user.blog.concat(savedBlog._id)
+      await user.save()
       response.status(201).json(savedBlog)
+
     }else{
       response.status(400).end()
     }
   })
   
-  blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+  blogRouter.delete('/:id', async (request, response, next) => {
+    try {
+      const decodedToken = jwt.verify(request.token, process.env.SECRET)
+      if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: "token missing" })
+      }
+  
+      const blog = await Blog.findById(request.params.id)
+  
+      if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        return response.status(204).end() 
+      } 
+      else {
+        return response
+          .status(401)
+          .json({error: "Blog can only be deleted by creating user"})
+      }
+    } catch (exception) {
+      next(exception)
+    }
+
   })
